feat(workbench): extract launchpad package
