+++ packages/workbench/launchpad/README.md
@@ -1,0 +1,3 @@
+# @tutti-os/workbench-launchpad
+
+Reusable Workbench launchpad UI, item model, and Dock entry helpers.
