+++ packages/workbench/launchpad/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../packages/configs/tsconfig/base.json",
+  "compilerOptions": {
+    "allowImportingTsExtensions": true,
+    "jsx": "react-jsx",
+    "types": ["node", "react", "react-dom"]
+  },
+  "include": ["src/**/*", "tsup.config.ts"]
+}
