+++ packages/workbench/surface/src/styles/workbenchStyle.test.ts
@@ -267,18 +267,6 @@ test("dock overflow keeps scroll controls viewport-bound", () => {
   );
   assert.match(
     css,
-    /\.workspace-launchpad-dock-icon\s*{[^}]*--workspace-launchpad-dock-icon-bg:\s*var\(--transparency-block\);[^}]*--workspace-launchpad-dock-icon-border:\s*var\(--line-1\);[^}]*background:\s*var\(--workspace-launchpad-dock-icon-bg\);/s
-  );
-  assert.match(
-    css,
-    /\.desktop-dock__slot\[data-wallpaper-tone="dark"\]\s+\.workspace-launchpad-dock-icon\s*{[^}]*--workspace-launchpad-dock-icon-bg:\s*rgb\(255 255 255 \/ 18%\);[^}]*--workspace-launchpad-dock-icon-border:\s*rgb\(255 255 255 \/ 15%\);/s
-  );
-  assert.match(
-    css,
-    /\.desktop-dock__slot\[data-wallpaper-tone="light"\]\s+\.workspace-launchpad-dock-icon\s*{[^}]*--workspace-launchpad-dock-icon-bg:\s*rgb\(0 0 0 \/ 10%\);[^}]*--workspace-launchpad-dock-icon-border:\s*rgb\(0 0 0 \/ 10%\);/s
-  );
-  assert.match(
-    css,
     /\.desktop-dock\[data-scroll-overflow="true"\]:hover\s+\.desktop-dock__scroll-button:not\(:disabled\)/s
   );
   assert.match(
@@ -405,42 +393,5 @@ test("dock overflow keeps scroll controls viewport-bound", () => {
   assert.match(
     css,
     /\.desktop-dock\[data-dock-placement="left"\]\[data-scroll-forward="true"\]\s+\.desktop-dock__items\s*{[^}]*mask-image:\s*linear-gradient\(\s*to bottom,/s
-  );
-});
-
-test("launchpad top bar remains draggable without stealing controls", () => {
-  const css = readFileSync(resolve("src/styles/workbench.css"), "utf8");
-
-  assert.match(
-    css,
-    /\.workspace-launchpad-overlay__content\s*{[^}]*-webkit-app-region:\s*drag;/s
-  );
-  assert.match(
-    css,
-    /\.workspace-launchpad-overlay__topbar\s*{[^}]*-webkit-app-region:\s*drag;/s
-  );
-  assert.match(
-    css,
-    /\.workspace-launchpad-search\s*{[^}]*-webkit-app-region:\s*no-drag;/s
-  );
-  assert.match(
-    css,
-    /\.workspace-launchpad-grid-viewport\s*{[^}]*-webkit-app-region:\s*no-drag;/s
-  );
-  assert.match(
-    css,
-    /\.workspace-launchpad-grid-viewport\s*{[^}]*overflow:\s*visible;/s
-  );
-  assert.match(
-    css,
-    /\.workspace-launchpad-dock-hover-panel\s*{[^}]*--workspace-launchpad-hover-panel-shift:\s*0px;[^}]*left:\s*50%;[^}]*right:\s*auto;[^}]*max-width:\s*min\(240px, calc\(100vw - 32px\)\);[^}]*transform:\s*translateX\(\s*calc\(-50% \+ var\(--workspace-launchpad-hover-panel-shift\)\)\s*\)/s
-  );
-  assert.doesNotMatch(
-    css,
-    /workspace-launchpad-dock-hover-panel\[data-hover-panel-align=/
-  );
-  assert.match(
-    css,
-    /\.workspace-launchpad-pages\s*{[^}]*-webkit-app-region:\s*no-drag;/s
   );
 });
